Rename project from "dstcalc" to "pdst-calc". Update package structure to move core logic from `lib/src` to `db/src`.
