+++ db/src/README.md
@@ -1,7 +1,7 @@
-# lib/src Directory
+# db/src Directory
 
 ```
-───lib/src/
+───db/src/
     ├── __init__.py           # Package initialization and version
     ├── drug_database.py      # Drug database loading functions
     ├── dst_calc.py           # Core calculation functions
